Export functions and rename
- getAllConfigFiles -> getConfigFilesList
- getAllConfigPaths -> getConfigPathsList
